+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- [TD-3720] Update structure domain (with children)
 - [TD-3522] Support for StructureNote management with workflow
 - [TD-3552] Executable implementations